more moving helper functions
